file-operations: Use g_file_enumerator_get_child
As documented, this is a convenience method that's equivalent to our
existing code, assuming the directory at hand is the enumerator
container.

Also, this convenience method remains correct if the enumerator
container is different, which is going to be the case soon, as we
introduce support for retrying with admin:// URIs.