file-undo-operations: Set NULL dest for duplication
In src/nautilus-file-operations.c, the duplication operation is
identified within the 'copy' code paths by having a NULL destination
imply it's the same as the source parent dir.

For consistency, adopt the same paradigm in the undo stack.

Also, while we are at it, adopt autocleanup for some related strings.