pathbar: Avoid leaking stack pages
The templates submenu in the current location popover is recreated
every time the menu model is updated, but the old widgetry is never
destroyed.

This results in a memory leak and many warnings in the terminal output.

I couldn't find the root cause, after many investigations. However,
I've found that unsetting the model actually removes the old widgetry.
Let's do that as a workaround.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/1705