files-view: Clear :templates-menu if it should not be used
If the templates dir is set to Home, then it should not be used.
We set the priv->templates_present boolean to FALSE in that case,
but forget to set the :templates-menu property to NULL too.

Fix that corner case issue. Also, drop the useless boolean and
just check directly whether the menu is present (i.e. not NULL).