files-view: Only refresh Scripts menu if needed
Same rationale as for Templates.

But, in this case, we don't have a cache ready to use.

So, keep a cache of the menu model in the private instance data.
No need for a GObject property in this case.

Also, remove scripts_present boolean, now that we can directly check
whether the scripts_menu is present (i.e., not NULL).