operations-ui-manager: Refactor main context task completion
invoke_main_context_sync() wraps the boilerplate needed for the file
operations thread to trigger UI dialogs and wait for their feedback.

It assumes that the UI task is completed by the time the invoked
function returns. This assumption presumably stems from the expected
usage of gtk_dialog_run().

However, gtk_dialog_run() is gone in GTK 4, so we should stop making
that assumption. Let's introduce an explicit way to signal that the
task is completed, such that it can be used either before or after
the wrapped function returns.

This will enable us to drop gtk_dialog_run() later. It will also
help with https://gitlab.gnome.org/GNOME/nautilus/-/issues/1829