file: Add constant for minimum size to use
Currently, list view doesn't use thumbnails for its first zoom levels:
SMALL (16px) and STANDARD (32px). So, it does display thumbnails for
LARGE (48px) and LARGER (64px). On the other hand, the icon view uses
displays for all zoom levels, the smallest being 48px.

In short: thumbnails are displayed for zoom levels that request an
icon size of 48 logical pixels. So, the minumum size of thumbs is 48px.

This is implemented by checking that the current icon size is equal to
or lower than NAUTILUS_LIST_ICON_SIZE_STANDARD (32px).

However, we are going to need to check this condition in other places,
so it would be best to have a constant specifically for this purpose.

For the sake of naming this constant reasonably, let's use the minumum
size of 48px and update the condition checking accordingly.