general: Remove thumbnail frame
Shadows are present in design mockups, frame is not.

In GTK 4 we are going to be able to draw shadows in the new views using
CSS, which is going to be much better than transforming pixbufs.

Furthermore, this removes another use of libgd, which we want to stop
depending on as part of the GTK 3-to-4 migration.