properties-window: use FilterType instead of gboolean
The gboolean "is-folder" stored as GObject data, indicates
whether a permission combo box or a checkbox refers to folders or files.
For enhanced readability and performance this is replaced with the
FilterType enum. This allows to combine checking the property window
target_files for files or folders and clarifies passed parameters to
helper functions.