properties-window: use AdwComboRow for permissions
Replace the previous GtkComboBox widgets for permissions with
AdwComboRow. This requires changing the underlying model from a
GtkTreeModel with indexed columns, to a GListModel and a helper struct
NautilusPermissionEntry.

Additionally permission description strings are now centralized via the
permission_value_to_string helper function. This replaces the previous
approach of piecing together strings for non-default options, allowing
for simpler translations and shared strings.

Part of #1326