properties-window: detect changed user via hash
When creating the drop down list of owners, store a hash value of the
used user names. Use this hash value to detect if the users of the
system have changed and the list needs to be recreated.

This removes a dependency on GtkTreeModel, easing future adaption to
GListModel.

Part of #1326