Revert "general: Stop overriding GtkWidget.destroy()"
This reverts commit 1c83d6af8bb4890dfb279c75506fa9be2a66b6ee.

It did a little to much and this introduced at least one bug: when
going from a regular folder to other-locations:///, the current view
menu keeps displaying invalid actions (New Folder, Properties, etc.),
which, if activated, crash the application.

Let's revert first and do it right in the next commit.