view-icon-controller: Implement remaining vfuncs
The implementation of NautilusFilesView vtable was incomplete.

Only .end_loading() remains a no-op because it is only used in other
views for the clipboard, so it's best left alone for now, as GTK4 will
change how clipboard is handled anyway.

Also, .widget_to_file_operation_position() is ignored because that's
a vestige of manual sorting, which doesn't exist anymore since the
icons-on-desktop removal. It's to be removed along with the canvas.