view-icon-controller: Fix selection when scrolled
We add a click gesture to the scrolled window, not the view itself,
because we need to get background context menus on whitespace around
the flowbox (not only padding but bottom whitespace when the rows
don't fill the whole height).

Because of that, the selected item may not be the one which the user
had selected (e.g. if the view is scrolled away from the top).

Translate the coordinates properly to properly select the item where
the user had clicked. Do the same for the longpress gesture.