view-icon-item-ui: Draw shadow only if thumbnail available
We have been drawing shadow on files which can be thumbnailed.

This results in two bugs:
- we draw shadow around the placeholder icon while creating thumbnails;
- we don't draw shadow around thumbnails if the thumbnailer is missing.

The second case happens under flatpak because we currently don't have
access to thumbnailer programs from outside the sandbox.

Instead, we should just check that a thumbnail path is set, meaning
the file has got a thumbnail, which is what we need to know in order to
correctly decide whether to draw a shadow.