view-icon-controller: Open new tab with middle click
And also with double-click while holding SHIFT, matching old views.

The GtkFlowBox::child-activated signal is no longer going to be used
for activation on click, as it doesn't provide the event data we need,
namely button, modifiers, and number of clicks.

While we *can* get that information thanks to gtk_get_current_event(),
this function is going away with GTK 4, so let's not increase our
reliance on it any more.

Instead, implement a new activation helper for usage with gestures.