general: Use in-tree copy of GtkPlacesSidebar
GtkPlacesSidebar is a public GTK 3 widget, but private in GTK 4, so we
need to have the sidebar in our own codebase if we are to keep using it.

Extend the script already in use for the places view and use it to copy
the places sidebar and its dependencies and patch it as necessary.

Also, construct it from code, because this in-tree places sidebar cannot
be used in a GtkBuilder UI template.