places-view: Fix regression
Unlike NautilusFilesView, NautilusPlacesView is not a subclass of
GtkGrid, but instead it's a subclass of GtkBox.

Casting it to GTK_GRID fails at runtime, showing the Other Locations
view as empty.
This is a regression caused by e9ae1add9cccee61d10e724b3d7f3526df8ad32a

Fix it by treating it as a GtkBox.