toolbar: Use models for view and app menus
Building menus with regular widgets brings too much complexity. Also,
it makes porting to GTK4 harder because GtkPopoverMenu as we know it
and its GtkModelButtons are gone.

Let's use a GMenuModel-built GtkPopover instead. While we are at it,
also move Undo and Redo to the primary menu, where they belong.