properties-window: Add discoverable Forget action
The "Forget association" action is found in a single item context menu
in the Open with tab of Properties window.

Single item context menus are discouraged design and the feature is
not discoverable. Other actions have buttons.

Furthermore, this is implemented using two GTK features that are gone
in GTK 4: GtkAppChooserWidget::populate-popup and GtkMenuShell.

Just add a button instead. This is due a redesign later anyway.