toolbar: Drop edit section from primary menu
The layout of the edit buttons cannot be replicated using a model-bound
GtkPopover in GTK 3.

Also, action are duplicated in other menus and quite out of context in
this menu in my opinion.

Finally, removing these items opens space for Undo/Redo to be moved.

This commit is a preparation for the next one.