files-view: Make preview an action
Preview is activated when Space key is pressed. It has been implemented
in each view's key event handlers. Except the new grid view, where it
is part of the GtkFlowBox::child-activated handler; as this is not an
event handler, it requires calling gtk_get_current_event() in order
to know the pressed key and possible modifiers.

This is problematic because gtk_get_current_event() is gone in GTK 4.

Instead, let's simplify this whole thing by making an action and
setting an accelerator. This will trigger before the event handlers,
so there is no need for the event handlers to do anything about it.