pathbar: Refactor signals
NautilusPathBar::path-clicked is never emited.
NautilusPathBar::location-changed uses a parameter type that's gone in
GTK 4: GtkPlacesOpenFlags.

Generate a GType for our own NautilusOpenFlags enum type to be used as a
signal parameter, remove the unused signal and repurpose its signal
handler to handle the second signal (which no longer has a signature
compatible with the GtkPlacesSidebar one).