properties-window: use AdwActionRow and GtkSwitch for executable bit
For a simpler user interface represent the executable bit with a
GtkSwitch.

The previously used GtkCheckButton had an "inconsistent" state, which
was used when the executable bit was not set for all files or not set
for all of user/group/other. Since GtkSwitch does not have such a
special state, a separate GtkRevealer combined with a tooltip is used
to handle this case.

Part of #1326