toolbar: Don't subclass HdyHeaderBar
In GTK4 we are not going to be able to subclass headerbars.

But subclassing GtkWidget directly implies adding a lot of
boilerplate to behave as a container.

The easy way around that is going to be AdwBin. Use GtkBin for the
time being, but introduce a wraper for a missing method.