general: Don't subclass GtkInfoBar
It's going to be a final class one we switch to GTK 4.

In preparation, become bins containing an info bar.
Also, stop using gtk_info_bar_get_action_area(), which is gone in GTK 4
and we are using it only to set the box to horizontal orientation,
which is the default anyway.