floating-bar: Stop overriding widget measurement vfunc
We are doing it in order to account for the border and padding.

That's unnecessary, as GTK already accounts for border and padding
from CSS. Actually, duplicating this work has been just adding more
padding, resulting in a disproportionate padding at the start side.

Remove the unnecessary overrides and just add 2 pixels more padding
to preserve the visual height.