list-view: Merge did_not_drag() into its caller
This function has only one caller and it relies on the passed GdkEvent
for everything.

However, in GTK 4, we are going to depend less on GdkEvent and more on
the gesture itself and its signal handler parameters.

So, inline the function to make sure we have everything at ready for
when we switch to GTK 4.