list-view: Use coordinates passed by gesture
We have coordinates right there in the gesture ::release signal, but
are still getting coordinates from GdkEvent.

This is because gtk_tree_view_get_path_at_pos() requires bin_window
coordinates, which happened to be what the GdkEvent coordinates are
(because bin_window is the event's GdkWindow).

In order to use gesture coordinates, we need to translate them into
bin coordinates.