dnd: Stop using GtkMenu
This is the last usage of GtkMenu remaining in our own codebase.

DND is due a reimplementation in GTK 4 and this particular menu (the
one which asks the drop action if you hold Alt while dropping) uses
a nested main loop and grabs.

So, there is no need to aim for perfect here. Just reimplement it
the lazy way as a GtkPopover with buttons and call it a day.