batch-rename-dialog: Use GtkMenuButton proper
We have been using GtkToggleButtons which popovers are relative to,
adding signals to entangle their states. Basically reinventing
GtkMenuButton.

So, just use GtkMenuButtons, which also allows us to set the menu
model directly in the UI template as a property of the button.
Furthermore, this solution is compatible with GTK 4, unlike the
current GtkToggleButton + GtkPopover setup.