compress-dialog-controller: Drop signals from UI file
We have signals defined in the GtkBuilder UI XML file. However, as we
are not using it as a template (we are not subclassing GtkDialog), we
still need to manually bind the callback symbols and even explicitly
bind them with gtk_builder_connect_signals().

In GTK 4 gtk_builder_connect_signals() is gone and the alternative is
setting the signal data (in this case, 'self') before loading the
UI definition.

However, we need to load the UI definition before we create 'self',
because some widgets are requires as construct-only properties. Given
this circular dependency, doing it the GTK 4 way is impossibile.

Therefore, just connect signals in code and remove them from the UI
file. This is compatible with GTK 4. And even under GTK 3 its simpler!