icon-info: Stop using gtk_icon_theme_choose_icon_for_scale()
This non-GIcon API is gone in GTK 4.

The forward-compatible way is to lookup by GIcon. Actually, we already
have a GIcon anyway, and we are going the long way of getting its icon
names to pass to the non-GIcon API.

So, just use gtk_icon_theme_lookup_by_gicon_for_scale(). Also, remove
a lot of now-redundant code. This is going to help switching to GTK 4.