window-slot: Drop non-window usage of gtk_widget_destroy()
When switching to GTK 4 we are going to replace gtk_widget_destroy()
with gtk_window_destroy(). This requires we drop all usage of this
method on widgets other than GtkWindows.

For the content view, add a gtk_box_remove() which matches the
gtk_box_append() already present in this file.

For the query editor, there is no need to destroy it here, because
this is not the file where it's added to its parent. It's added in
the toolbar, and the toolbar already has a weak ref on the slot to
do this when the slot is destroyed. So, simply clear the reference
we have sunk in this file.