files-view: Stop setting custom filechooser filter
We set a custom filter to prevent selecting a folder that's part of the
selection as the destination to move that selection.

GTK 4 doesn't support custom filters, and the available options don't
fit this purpose.

Just remove it. It's safe: we popup an error dialog saying you can't
move a folder into iself. Of course, it would better not to allow the
user to make this mistake, but let's we can't have everything.