file-operations: Prevent overriding source on move
This had been fixed but only for the fallback implementation, as per
100781f806b922cf6b959ec663a196a47b6eb71e

As a result, when moving a file/folder to its current location, the
default implementation still tries g_file_move(), which obviously
fails with a conflict, then presents the user the conflict dialog.
Then:
  - for a file, if the user chooses "Replace", then g_file_move()
proceeds to replace the file with itself, which is... useless at best.
  - for a folder, if the user chooses "Merge", g_file_move() still
fails because it doesn't support merging, so we try the fallback
implementation which, finally, warns the user that this is not allowed.

So, the first case is useless and the second case asks the user if
they want to perform an action which is not allowed anyway.

This issue is reproducible using the "Move to..." action.

So, copy the fix to the g_file_move() path. It's probably an oversight
of the original fix that it was done for only one path.