files-view: Stop connecting to GtkFileChooser::current-folder-changed
The signal is gone in GTK 4.

We have been using it to disable the action button if the destination
is a descendant of selection, which is not allowed for copy/move.

Ideally, we should reimplement this feature as a mode in our main
window, instead of using a file chooser. Until then, we can rely on
our file operations code to impede and warn about a copy-move operation.