properties-window: use GtkListBox for basic entries
Convert property entries in the basic tab from a grid form into
multiple GtkListBoxes. Display name and type centered as a title.
Timepoint entries, as well as the Trash-related entries are special
cased, in that they are combined into a single row with labels
arranged in a GtkBox.

The icon picker and the volume pie chart remain unchanged, with only
some placeholders used for spacing removed.

Part of #1326