pathbar: Stop overriding GtkWidget.style_updated() and .screen_changed()
Both vfuncs are gone in GTK 4.

They were used for the sake of updating icons live the icon theme is
changed. I don't see any need for that. If GTK doesn't update the icons
automatically, the icon is updated next time the pathbar is updated for
other reasons (such as switching locations).