notebook: Translate coordinates to find clicked tab
In GTK3, the tab label allocation was given in coordinates which were
relative to the parent widget's allocation.

In GTK4 the allocation is for the widget's own coordinates instead.

Translate the event coordinates accordingly to make this logic work
again. Also, test for the lower limit in addition to the upper limit.