window: Attach tab context menu to the tab
Due to GtkPopover now being a child of the window, instead of setting an
attach widget we need to pass it the allocation of the tab, with the
origin coordinates translated to the window coordinate space.

Also use double instead of int when appropriate and replace GdkEvent
usage with event controller method to get modifier status.

Closes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2101