view-icon-controller: Add style class
Our custom CSS relies on `flowbox` and `flowboxchild`. But GtkFlowBox
is about to be replaced with GtkGridView. Also, it may happen that
we use either widget in other contexts in this app in the future, so
it's best to be more specific on where we want the styles to apply.