icon-controller: Add gestures to children
GtkFlowBox allows us to get a child by its coordinates, but this will
not be possible with GtkGridView.

Instead, let's add gestures to each children such that we can get the
clicked widget using gtk_event_controller_get_widget(). This requires
using the bubble phase such that children can capture the event before
the parent, and translating event coordinates to the view ones. As we
can't get the position from the event widget, also introduce a method to
get the model item from the widget.

While here, rename an outlier callback symbol and remove a GTK3 #if.