view-icon-item-ui: Fix labels sizing
We want all items to have the same width and for this we have been
setting GtkLabel:max-width-characters to 0, such that labels take
grow only until the width defined by the icon container.

Unfortunately, doing this for wrapping labels in a dynamic grid causes
allocation warnings and visual glytches (labels overflowing the height
of their container), which I assume are due to the labels reporting a
minumum height for a width which goes beyong the minimum width set on
the icon container.

So, instead of setting Gtk:Label:max-width-characters to 0, wrap the
whole thing in a horizontal AdwClamp with its :maximum-size set to 0.
This forces the labels to report a height which corresponds to the
actual minimum width set by the icon container.

Also, remove some meaningless properties from the UI file.