properties-window: use level bar to visualize disk
For filesystems replace the previous pie chart widget with a GtkLevelBar
and show it on the top. Reserved space is now included under 'used'
without a separate representation. Total, used and free disk space are
indicated as a legend under the bar, with used and free being
color-mapped through a respectively colored dot.

The 'Open in Disks' button is replaced with an inline row action,
indicated by a newly added 'external-link-symbolic' icon.

Filesystems no longer display 'Folder' as a type, but instead the actual
filesystem type.

Part of #1326