view-icon-item-ui: Fix icon sizing
We want the top of the labels to align vertically among all items. For
this we set a minimum height and width for the container of the icon.

However, then a thumbnail is taller than it is wide, the picture makes
its container grow taller than the minimum height, despite the picture
never actually using being taller than the minimum height.

As in the previous commit, as a workaround, wrap the image in AdwClamp,
such that the height for width computation doesn't account for more than
height than necessary.