placessidebar: Clarify "Remove Bookmark" in the bookmark popup menu
Currently, in the bookmark popup menu, the option to remove the bookmark is named "Remove".

However, this may cause confusion to a new user/a user with weak knowledge about how the bookmark section whether the "Remove" option in the popup menu means removing the entry or removing the contents of the folder in the bookmark.

To fix this, the option "Remove" has to be changed to "Remove from Bookmarks".

Closes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2125