trash-bar: Remove trash infobar
Currently trash infobar contains actions for restoring files and
emptying trash.

This is misusing of the GtkInfoBar, as it should only be used to present
some kind of status message, and having so many different buttons for
actions is uncommon and resulted in https://gitlab.gnome.org/GNOME/nautilus/-/issues/2096

To fix that, remove the trash infobar altogether, with intention of
rolling it into special-location-bar, as it will only have 1 button.