floating-bar: Fix hiding on hover
Since switching to GTK4 the floating bar no longer automatically hides
while hovering the pointer over it.

This is because in GTK 4 entering the child area (floating bar in our
case) isn't treated as leaving the parent, so fix this situation by
moving the relevant code to the on_event_controller_motion_motion() that
is called on every motion inside the parent of the floating bar.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2100