floating-bar: Check x coordinate when hiding
Ever since the start of hiding the floating bar on hover, it has been
possible to keep the bar hidden even when not hovering on its place by
moving the cursor to the left.

This is because only the y coordinate has been checked in appropriate
functions, so check for the x coordinate as well.