canvas-item: Fix black corner glitch in selected thumbnails
The  canvas item style includes a 5px border-radius for the rectangle
around labels when the item is selected.

However, this is also affecting the icon, causing black artifacts in
the corners of selected thumbnails and custom icons. Previously, this
wasn't visible for thumbnails because they used to have rounded frames.

Explicitly set the 5px for the box label alone and ensure the icons
(before applying the shadow) have sharp corners.