files-view : added file extensions to "New Document" submenu entries
Currently, while creating a new file using the "New Document" Section
the file extensions of the template files are not shown. This may lead
to confusion if there are files in the Templates directory with the
same basename but with different extensions. To fix this, the full
filename is now shown.

Closes #2167