css: Add backdrop style for pathbar
The difference between @borders and @unfocused_borders colors is barely
noticeable with the regular stylesheet, but marked with high contrast.
The same is true for @theme_bg_color vs. @theme_unfocused_bg_color.

So, under the Adwaita spreadsheet, this change is barely noticeable,
despite being correct and indeed a bugfix.

The main purpose of this patch is to improve HighContrast support.