pathbar: Fix button list order regressions
The pathbar code makes the assumption that the first GList link in the
list of path buttons corresponds to the current location and that each
subsequent link is corresponds to the parent location of the previous
one.

This assumption relied on reversing of the list of buttons, which I've
broken in commit eab10931491aba1c2d92ec716d8c9de91660c85d

Fix the misbehaviours by restoring the list reversion.