pathbar: Fix button list order regressions
The pathbar code makes the assumption that the first GList link in the
list of path buttons corresponds to the current location and that each
subsequent link is corresponds to the parent location of the previous
one.

This assumption relied on reversing of the list of buttons, which I've
broken in commit eab10931

Fix the misbehaviours by restoring the list reversion.

(Forward porting 67a3fe05ad2e0994101b0a668af1c6c321f0ca47)