general: add keyboard shortcut for starring
Currently, there is no keyboard shortcut for starring/adding files as
favorites.

However, for a lot of users who prefer using the keyboard instead of a mouse,
using the popover/context menu is more cumbersome.

To fix this, a new keyboard shortcut control+s is added for starring files.
This has also been updated on the shortcuts help window.

Fixes #2194