files-view: Drop non-portal wallpaper setting code
The "Set as Wallpaper" action has been originally introduced in a time
when the Settings background panel did not provide the ability to pick
a file to set as wallpaper. The action in the Files app was, therefore,
essential.

Nowadays, the Settings Appearance panel provides a button to pick any
picture file, so this action is no longer essential. Furthermore, we
can optionally use the Wallpaper portal instead of managing the setting
ourselves.

Therefore, there is no need for us to keep the maintenance burden of a
setting which is so often out of sync with other implementations.