css: Fine-tune floating bar visuals
Currently, the floating status bar has a custom border, which needs to
be changed depending on text direction, not to conflict with window border.

But this border looks bad when overlaying colorful content.

To fix this, instead of using border, we need to use box-shadow with 1px 
spread radius. Thanks to the clipping at the edge of the window, there is
no need to worry about text direction.

https://gitlab.gnome.org/GNOME/nautilus/-/issues/2099