pathbar: Avoid excessive minimum label width
We set a minimum width in characters for the labels in the pathbar not 
to collapse into a single ellipsis. Labels under this threshold are
excluded because in that case we just set them not to ellipsize and let
GtkLabel take care of the sizing.

However, the width in characters metric fails for variable-width fonts, 
such that in some cases we still set a wider label width than what the 
label text occupies.

So, raise the ellipsizing threshold further, and do it proportionally.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2202

(cherry picked from commit aa4a6a963d6bd0d27ce47bcecbd8160794586e51)