pathbar: Fix crash when right clicking
button_data_free() was destroying the button_menu_popover because it
was set as a child of the button previously.

This means that the next time one right clicks a parent folder, we
use a pointer to a destroyed object, crashing the application.

Reparent to the pathbar before the button is destroyed.