enums: Redefine icon size options
The new grid view introduces a new 256px icon size option, optimized
for thumbnails. This was made possible because, unlike the old canvas,
we have more control over the padding and can keep it low.

However, with it, we have 5 size options now, which a little too many;
in particular, the 64px size doesn't differ significantly enough from
the 48px nor the 96px option now (the old canvas view this difference
was more marked because padding increased in proportion).

Similarly, the 48px option in the new list view is not significantly
different from 32px and 64px. In the past, 32px didn't show thumbnails
because the frames didn't allow the thumbnails to use the whole 32px;
this would mark a difference vs. 48px. With the removal of thumbnail
frames, this distinction is also gone.

The 128px option in grid view is also not significantly larger than
96px and yet significantly smaller than 256px.

Drop the superfluous sizes, replace 128px size with 168px size for a
geometric progression and rename the enum symbols.