places: Align popover menu with mouse position
The popover menu always pops-up in the center of each
row regardless of where the mouse cursor is currently positioned.

Make the popover popup at the pointer click position. If the popover
is triggered by the keyboard (i.e. SHIFT+F10), then align it with the
start of the row.