issue_templates: Add a default template
Add a default template that will be displayed by default when opening a
new issue. This encourages usage of the template selection, that users
not familiar with reporting issues might otherwise miss.

Users familiar with reporting issues will have the option to clear the
text and file an issue from scratch.