placessidebar: Point to row instead of reparenting
In GTK 3 we used to move the popover around using set_relative_to();
this is gone in GTK 4 and the apparent direct replacement is setting
the target widget as the new parent.

But this requires a lot of careful handling least the popover become
orfan, which gets us ready to crash at any moment.

Since we only care about positioning the popover relative to a row,
let's use the set_pointing_to() instead of reparenting. Now, the
sidebar is always the parent.