file: Fix remote filesystem check
Files are flagged as remote in file->details->filesystem_remote
through directory-async as folders are browsed.
nautilus_file_get_filesystem_remote() incorrectly returns FALSE for
child directories that have not been processed by directory-async
on a remote filesystem.  This causes the "Count Number of Files in
Folders" options to not work correctly when set to "On This Computer
Only".  Rather than unnecessarily process all of the child folders,
check whether filesystem_info_is_up_to_date==TRUE.  If not,
check whether the parent is remote instead.

This issue was introduced in 99f55748ef69a68df86b9a039998ee576d59b898.
Previously the prefix of the URI was checked to match a hardcoded list.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2035