files-view: Drop non-portal wallpaper setting code
When the Wallpaper portal is not available, the "Set as Wallpaper" action falls back into the old implementation.

This fallback offers a subpar user experience:

  - there is no preview: the user might find a weirdly resized image at the desktop at a later moment;
  - there may be interruptions by file conflict dialogs, if setting a picture which had already been set this way in the past;
  - it's often out of sync with other implementations (a recent example being the new dark mode wallpaper); this is also aggravates the maintenance burden.

This old implementation "Set as Wallpaper" action has been originally introduced in a time when portals didn't exist yet and the Settings background panel did not provide the ability to pick a file to set as wallpaper. The action in the Files app was, therefore, essential.

Nowadays, the Settings Appearance panel provides a button to pick any picture file, so this action is no longer essential. As such, we can just drop the old implementation and hide the action if libportal is not available.

Closes https://gitlab.gnome.org/GNOME/nautilus/-/issues/1677