mime-actions: Close broken link message dialog
We used to rely on gtk_dialog_run(), which would allow us to block the
execution of the current function.

But it's gone in GTK 4, so we've ported to the ::response signal.
8af45e55fcc347b05c4376128df5f248732d07d8

However, a 'goto` which relied on the blocking behavior was kept by oversight,
such that the ::response signal wasn't connected and, as a result, the dialog
didn't close on response.

Close https://gitlab.gnome.org/GNOME/nautilus/-/issues/2265