view-icon-controller: Don't share click handlers
In the past, there used to be a single click event handler for the
whole view. The coordinates would be used to detect whether the click
happened on an item or the background.

Nowadays, with a list model widget, we can't use the coordinates to
find clicked items, so we have instead:

    - 1 click gesture for the whole view;
    - multiple click gestures for each item widget.

Despite multiple gestures, their signal handlers still share a
common path. This makes for long callback definitions and loss of
clarity.

Split the click callbacks into item callback from view callback.