view-icon-controller: Deny background actions when item is clicked
When the view background is clicked, the current selection is cleared.

However, we have no simple way to get events from background only.
So, we must rely on the click event controller for the whole view area.

As a result, when an item is clicked, the selection is cleared, which
is undesirable. By dumb luck the clicked item still gets selected in
the end, but when dealing with multiple selections and, soon, drags,
the bug shows its ugly face.

The simple solution is keeping a boolean state during clicks on items,
based on patch by Corey Berla <corey@berla.me>.