view-icon-controller: Select single item on specific cases
We've been changing the selection to the clicked item (if unselected) in
all cases except when SHIFT or CTRL are pressed, because the default
GtkGridView behavior does it only for primary click and only after
release.

But this causes duplicate selection events in some cases, interferes
with drag-and-drop (work in progress in another branch) and misses some
corner cases (e.g. middle click while holding CTRL).

Instead, factor out that block into a helper function to be called in
the specific situations where we need it:

    - before activating with middle click;
    - before opening context menu;
    - before activating with single click.

Double-click doesn't need it because the item is selected by the release
handler of the first click.

Single-click activation enforces single-item selection for consistency
with double-click activation. Multi-item activation still possible with
middle click, accelerators and context menu.