Drop libportal option
This is a follow up from https://gitlab.gnome.org/GNOME/nautilus/-/commit/6af38c29d

As a result of that commit, it's not possible to set a picture as desktop wallpaper
from Nautilus unless libportal is used. Since libportal is generally available,
it's no longer a useful option to not use libportal.