Files-view: Rely on GTK default item activation shortcuts
Currently it is not possible to activate any UI element with Enter on
the keyboard, because it always activates either the selected file in
the files view, or it activates the secondary menu next to the path.
This is because a) NautilusFilesView registers a global application
accelerator for Enter/Return, to activate the selected files, b) it
should really not have registered a custom shortcut for that to begin
with, instead it should use the grid view or list view's
"(row-)activate(d)" signals, which is already the case for the later.

Hence drops this Enter/Return global accelerator, and hooks up the
NautilusIconViewController to its grid view's "activated" signal.

Fixes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2225