sidebarrow: Hold reference to nautilus file
When dragging to the sidebar, we check whether the source and dest
files are on the same file system.  Since the source file is already
loaded in the files view, we have a nautilus file with preloaded
filesystem info.  The dest file, however, doesn't have file system
info available.  Rather than checking for file system over and over
as we drag accross the sidebar, keep a reference of the Nautilus File.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/1781