toolbar: Move "Reload"/"Stop" actions into folder menu
Although we are technically reloading the view, that's an implementation detail.

From a user point of view, we are reloading the current location. The main point
of reference here can be web browsers, which also have got a Reload action in 
the page context menu.

So, move the action to its due place.

Closes #1860