nautilus-window-slot: Open with correct view after navigating to other-locations
When navigating to Other Locations with "Open in New Tab" or
"Open in New Window",  you are always presented with the icon view
because view_mode_before_places was never initialized.  An uninitialized
gint == 0 (and NAUTILUS_VIEW_GRID_ID == 0).

Initialize view_mode_before_places in init to NAUTILUS_VIEW_INVALID_ID
so that it will follow the default behavior.  For best practices,
change NAUTILUS_VIEW_INVALID_ID to 0 (which would have also prevented
this issue) and NAUTILUS_VIEW_GRID_ID to 2 (matching the keyboard
shortcut).

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2281