gtkplacesview: Name root directory after OS
Having a "Computer" palce inside "On This Computer" section does not make
sense and sounds strange. Also, if someone wants to save files in their 
computer, the "Computer" location might sound like the right place, and 
it's not.

Sometimes the root directory is not even the root of a disk or partition, 
but some special mount (e.g. under ostree).

The root directory is directly related to the operating system, so let's
name it after the operating system. This helps convey the idea that making 
changes there is going to affect how the system works and possible break it.

Closes https://gitlab.gnome.org/GNOME/nautilus/-/issues/2250