list-base: Fix extending selection after going to parent
When going to parent, the previous location is selected and focused,
but using Shift + arrow keys to extend selection has unexpected
results: it assumes the last item is the start of the selection
range.

There is probably a GtkListBase bug underneath, but let's add a hack
to fix the symptom for the time being.

Closes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/2294