list-base: Temporary workaround for rubberband during dnd
While performing a DnD operation, the rubberband begins and is not
correctly cancelled in GTK causing unexpected behavior. This is
fixed in: https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/4831

Until the issue is fixed in GTK, use temporary workaround to disable
rubberbanding on item click and reenable rubberbanding on
stop/release or view change.